+++ tests/testthat/test-cluster.R
@@ -12,7 +12,7 @@ test_that("test parallel collisions", {
 
     # make archivist repository
     if (!file.exists(CacheDBFile(tmpdir))) {
-      if (getOption("reproducible.newAlgo", TRUE))
+      if (getOption("reproducible.useDBI", TRUE))
         createCache(tmpdir)
       else
         archivist::createLocalRepo(tmpdir)